Updates to the Examples
Minor fix to the Base URL passed into the proxy, it was including the
"/v2" in the path when these were already included in the path defined
in the com.hashnot.etsy interfaces.

Also filled out the example a bit to show getting the response and 
printing its contents to a logger.

Added an additional Example setting up Oauth