Account for responses that only return one object and not a list
The Response object has a getTypes that returns a List<T>, but some
of the responses from Etsy are not a list of things. Using the Response
type was causing a JSON parsing error.

Looked like the DictionaryReponse might be the correct type for handling
a single response as it had all the other fields of a Response (count,
type, pagination), so added a getType that just returns a T.